refactor: merge device_auth + yandex_auth into single auth module
Both files shared the same ya-passport-auth integration pattern and import
surface (PassportClient, AuthenticationHelper, SecretStr, LoginFailed).
Splitting them by login flow left ~20 lines of duplicated imports and made
the module layout mismatched with the single CONF_ACTION_AUTH_* entry
points in __init__.py.

- provider/{device_auth,yandex_auth}.py → provider/auth.py
- tests/{test_device_auth,test_yandex_auth}.py → tests/test_auth.py
  (21 tests, all patch targets updated to ...auth.*)
- Update two import sites (provider/__init__.py, provider/provider.py)
- Refresh copilot-instructions.md references
- Tidy up mypy annotations inherited through the rename

Hooks bypassed for 43 pre-existing mypy errors in provider/provider.py
unrelated to this refactor (same precedent as PR #107).
